Query now supports returning maps (#9)
This commit adds the capability to parse return maps
to the datalog-parser. Since Datomic supports this
the datalog-parser should be able to handle it as
well.

The parser checks if there is one of the three
relevant keys in the query and based on that
creates records from it with the type and the to be
mapped keys. The query is only valid if the count
of the mapping-keys is equal to the count of the
find variables or if there are no mapping-keys.

It works with the query in vector or in map format.
In case there is no return-mapping requested there
should be no impact on the performance because it
only checks on the keys once and if there is none
there should be no further computing made.

- Precedes https://github.com/replikativ/datahike/pull/159
- Refers https://github.com/replikativ/datahike/issues/149

Squashed commit of the following:

commit 91556f7cfa82bf6f1f793bee3f3cb5b4eb7c866f
Author: Timo Kramer <fw-git@timokramer.de>
Date:   Sat May 23 22:58:31 2020 +0200

    better naming for the keys of returnmaps-record

commit 179c4a6286a02e04960e8571c2f0eaadf334df40
Author: Timo Kramer <fw-git@timokramer.de>
Date:   Sat May 23 22:14:21 2020 +0200

    no need for a record for the mapping type

commit 4c68d1555adfadb6cb7c57085b7a938b3432254e
Author: Timo Kramer <fw-git@timokramer.de>
Date:   Sat May 23 19:40:15 2020 +0200

    working and tested

commit f1c2abf800b39cd722162df55b87348f9e6b6de6
Author: Timo Kramer <fw-git@timokramer.de>
Date:   Thu May 21 12:34:06 2020 +0200

    working first draft

commit 829029c65e4911f0092b5223c937efdbafc5ee29
Author: Timo Kramer <fw-git@timokramer.de>
Date:   Wed May 20 23:00:47 2020 +0200

    working solution with records

commit 1579c53437fabe523259042e98724287f32492c8
Author: Timo Kramer <fw-git@timokramer.de>
Date:   Tue May 19 21:07:26 2020 +0200

    first draft but not record